server/parser/requirements.txt
llama-index-core
llama-parse
llama-index-readers-file
python-dotenv